fix(locate): resolve by source position, not parser block order
Seven independent reviewers (six persona lenses plus a cross-model pass) found
one premise the plan asserted and nobody checked: document.blocks() is not in
source order. comrak emits every footnote definition after the blocks that
reference it, so the list is two ascending runs.

P0 — block_at binary-searched that list. In a document with a footnote
definition followed by more content, roughly half the byte offsets resolved to
block: None, which the docs define as "you clicked a blank line". A click on a
table five lines below a footnote returned no block, no row, no etag. Measured
36 of 64 offsets wrong in a 64-byte document, 3522 across 200 generated
layouts. It now scans; the spans are disjoint even when unordered, and the scan
costs nothing next to the SectionIndex this call already rebuilds.

P0 — build_preamble took its bounds from the first and last block in vec order.
On a footnote-first document that yields byte_start > byte_end and the next
slice panics, so every locate call on such a document panicked against a doc
comment promising one error case. Bounds are now min/max over source position.
Pre-existing (md outline panics identically on master) but newly reachable.

P1 — section_for_block read block_indices, built by walking blocks in index
order, so a footnote definition physically under one heading was attributed to
the next. The P0 masked it; fixing the P0 unmasked it. Both lookups now key off
source position, and locate uses the byte lookup in every case.

P1 — locate_line resolved from a line's first byte, but up to three spaces of
indentation are legal and belong to no block, so locate_line(doc, 1) on
"   para\n" found nothing. It now skips indentation to the first content byte,
and a blank line still resolves from its own start.

P1 — a located section could not drive a section edit: SectionIndex::new built
every heading selector with occurrence: None, so feeding it back to resolve
failed DuplicateHeading on repeated headings. The occurrence now travels with
the entry.

P2 — the preamble's span stops at its last block while a heading section's runs
to the next heading, so the blank line above the first heading belonged to no
section and returned section: None, which the docs define as "frontmatter".
The index now carries the preamble's region separately from its span, leaving
`md section :preamble` bytes and etag untouched.

P2 — locate_line(doc, line_count) returned an out-of-range byte error for an
in-range line. It returns Ok with every field None: that line holds nothing,
and `for line in 1..=line_count()` is the loop this function exists to serve.

P2 — a click on a table row's terminating newline lost the row, because
containment reused the newline-excluded span the mutation splices over.

P2 — table_etag recomputed the fingerprint instead of reusing the block
record's, a second copy of a formula that had already been guessed wrong once.

Test quality: located_task_loc_round_trips asserted only that the located loc
was *some* known loc, so an implementation returning the outermost task would
have passed; it now pins the exact loc, span and etag. The agreement test ran
on a fixture with no footnotes — the one shape that breaks it — and now runs on
tests/fixtures/footnote_midbody.md.

Also: Located gains PartialEq/Eq/Default and both new types gain
non_exhaustive; ByteOffsetOutOfRange.source_len is u32 like every sibling; the
ADR, README and CLAUDE.md drop the false "sorted" claim, stop naming
replace-section as a mutation these etags feed, disclose that the guard is
content-addressed while the located target is identity-addressed, and carry
measured cost numbers instead of a bare O(blocks).

670 tests pass; clippy warning count is unchanged at 74.

+++ src/core_error.rs
@@ -38,7 +38,7 @@ pub enum CoreError {
     },
     ByteOffsetOutOfRange {
         byte_offset: u32,
-        source_len: usize,
+        source_len: u32,
     },
     LineOutOfRange {
         line: u32,
